django/requirements.txt
apig-wsgi==2.18.0
asgiref==3.8.1
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-filter==24.3
django-ses==4.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
